backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
pydicom==2.4.3
pillow==10.1.0
numpy==1.25.2
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0
python-dateutil==2.8.2
uuid==1.30
typing-extensions==4.8.0
aioredis==2.0.1
loguru==0.7.2
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6
gunicorn==21.2.0